LPD-77080 move BaseTemplateUpgradeProcess to com.liferay.portal.kernel.upgrade so that it can be called from different unrelated modules.
